Move the bulk of test_psa_compliance.py to a module
This will allow consuming branches to each have their executable entry point.
